Move extensions to extensions class
